requirements.txt
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pymongo==4.2.0
Werkzeug==2.2.2
zipp==3.8.1
